requirements.txt
# Backend runtime — Franken-Fit hackathon build
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# Tier A integrations
google-genai>=1.0.0          # Gemini Vision + TTS + text (modern SDK)
fal-client>=0.5.0            # FLUX.2 multi-reference edit + image-to-video
httpx>=0.27.0                # Tavily, Pioneer, eBay
Pillow>=10.0.0               # Image preprocessing for fal upload
